requirements.txt
aniso8601==4.0.1
backports.shutil-get-terminal-size==1.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==1.7.2
decorator==4.3.0
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.6
flask-mongoengine==0.9.5
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
idna==2.7
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.16.1
mongoengine==0.15.3
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.2
pytz==2018.7
PyYAML==3.13
requests==2.20.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.5.3.0
urllib3==1.24
uWSGI==2.0.17.1
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
